feat: change int to long in cache to be compliant with DB

+++ order-payment/src/main/java/com/codecool/order_payment/dto/NewOrderDTO.java
@@ -3,5 +3,5 @@ package com.codecool.order_payment.dto;
 import java.util.Map;
 import java.util.UUID;
 
-public record NewOrderDTO(UUID userID, Map<UUID, Integer> items) {
+public record NewOrderDTO(UUID userID, Map<UUID, Long> items) {
 }
